separate entrypoint-7702 tests into a separate test file
